requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
python-dotenv>=1.0.0
aiohttp-socks>=0.8.4
python-socks[asyncio]>=2.4.3
psutil>=5.9.0
GPUtil>=1.4.0
WMI>=1.5.1; sys_platform == 'win32'
Pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
